pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "job-scraper"
version = "0.1.0"
description = "Security job discovery via local SearXNG instance"
requires-python = ">=3.10"
dependencies = [
    "requests",
    "pydantic>=2.0",
    "pyyaml",
    "rich",
    "typer",
]

[project.scripts]
job-scraper = "job_scraper.__main__:app"

[tool.setuptools.packages.find]
include = ["job_scraper*"]
